Add a script to ensure commits referenced in countries.json are valid
The countries.json file provides URLs to Popolo JSON files and CSV files
which include the full object name of a commit that contains the
up-to-date version of that file.  It's very easy, however, to forget to
rebuild the countries.json file after updating the data for a countries,
or rewriting the history of a branch in which you made an update.

This script should detect both of those errors, emit a warning and exit
in error if any problems were found.

If this is useful, it could be added to checks run by Travis for each
pull request.

(This is just a proof-of-concept to see if people think it'd be
helpful, so I haven't written any tests yet.)